backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.1
langchain-google-genai==2.0.0
langchain-core==0.3.0
pydantic==2.9.0
tenacity==8.5.0
PyPDF2==3.0.1
fpdf2==2.8.1
